+++ packages/react/src/chains/evm/testnets/index.ts
@@ -1,6 +1,7 @@
 export * from './abstractSepolia.js';
 export * from './beraChain.testnet.js';
 export * from './firechain.testnet.js';
+export * from './monad.testnet.js';
 export * from './oasisSapphire.testnet.js';
 export * from './pentagon.testnet.js';
 export * from './soneium.testnet.js';